se comentan los servicios y controller, se logra persistir

+++ src/main/kotlin/ar/edu/unsam/phm/uberto/dto/LoginDTO.kt
@@ -12,11 +12,6 @@ data class LoginResponse(
     var jwt: Int
 )
 data class LoginDTO(
-    var id: Int,
+    var id: Long,
     var rol: Role
 )
-
-//fun User.toLoginDTO() = LoginDTO(
-//    id = id,
-//    rol = role
-//)
